+++ R/get_sv.R
@@ -20,7 +20,7 @@
 #' # svclass is optional
 #' rs2 <- read_sv_as_rs(sv[, setdiff(colnames(sv), "svclass")])
 #' identical(rs, rs2)
-#' \donttest{
+#' \dontrun{
 #' tally_rs <- sig_tally(rs)
 #' }
 #' @testexamples
@@ -146,7 +146,8 @@ getDists <- function(chrom1, pos1, chrom2, pos2, doPCF = FALSE) {
     if (!doPCF) {
       return(forCN[, 3])
     }
-    return(list(info = forCN, seg = copynumber::pcf(forCN, gamma = 25, kmin = 10)))
+    pcf = "copynumber"%:::%"pcf"
+    return(list(info = forCN, seg = pcf(forCN, gamma = 25, kmin = 10)))
   }, simplify = FALSE)
   return(dists)
 }
@@ -210,6 +211,10 @@ getType <- function(sv_profiles) {
 
 # get features list -------------------------------------------------------
 get_features_sv <- function(sv_data) {
+  if (requireNamespace("copynumber")) {
+    stop("'copynumber' package is not available. Please install it.")
+  }
+
   field <- c(
     "clustered",
     "type",