+++ NEWS.md
@@ -1,5 +1,6 @@
 # sigminer 2.2.0
 
+- Removed package 'copynumber' from suggests filed.
 - Supported Ziyu Tao et al approach for copy number segment classification. 
 - Supported ce11 genome in `read_vcf()`.
 - Added `read_maf_minimal()` to support a minimal MAF-like data as input.